rename ColoredRadio -> ColoredRadioButton
