+++ packages/ui/src/lib/coloredCheckbox/Introduction.mdx
@@ -6,4 +6,4 @@ import { Meta } from '@storybook/blocks';
 
 This provides a (square) colored checkbox component that can act as both a categorical color legend, and a control for which categories of things are displayed on a map or visualization.
 
-It should be used in cases where **multiple** items from a list of options can be selected simultaneously; if only **one** item can be selected at a time, then a round [ColoredRadio](./?path=/docs/ui-coloredradio--docs) should be used instead..
+It should be used in cases where **multiple** items from a list of options can be selected simultaneously; if only **one** item can be selected at a time, then a round [ColoredRadioButton](./?path=/docs/ui-coloredradiobutton--docs) should be used instead.
