--- packages/ui/src/lib/coloredRadio/Introduction.mdx
@@ -1,9 +1,0 @@
-import { Meta } from '@storybook/blocks';
-
-<Meta title="Ui/ColoredRadio" />
-
-# Colored radio button
-
-This provides a (round) colored radio-button component that can act as both a categorical color legend, and a control for which categories of things are displayed on a map or visualization.
-
-It should be used in cases where **no more that one** item from a list of options can be selected; if **multiple** items can be selected simultaneously, then a square [ColoredCheckbox](./?path=/docs/ui-coloredcheckbox--docs) should be used instead..
