+++ packages/ui/src/lib/coloredRadioButton/ColoredRadioButtonGroup.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import ColoredRadio from "./ColoredRadio.svelte";
+    import ColoredRadioButton from "./ColoredRadioButton.svelte";
     export let selectedId: string;
     export let name: string;
     export let options: { id: string, label: string }[] = [];
@@ -12,7 +12,7 @@
     {/if}
 
     {#each options as option}
-        <ColoredRadio id={option.id} label={option.label} color={option.color} bind:selectedId={selectedId}
+        <ColoredRadioButton id={option.id} label={option.label} color={option.color} bind:selectedId={selectedId}
                       name={name}/>
     {/each}
 </div>
